requirements.txt
# LangChain Core
langchain
langchain-core
langchain_community
langchain-mcp-adapters
langgraph
mcp

# RAG
chromadb
tiktoken

# OpenAI Integration
langchain-openai
openai

# Environment Variable Management
python-dotenv
redis
